feat : DDD 기준 적용
