+++ src/main/java/com/cho/polio/presentation/security/config/SecurityConfig.java
@@ -1,0 +1,46 @@
+package com.cho.polio.presentation.security.config;
+
+import com.cho.polio.presentation.security.authroization.PermissionRuleUriMapper;
+import com.cho.polio.presentation.security.converter.KeycloakJwtAuthenticationConverter;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configurers.ExceptionHandlingConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+
+
+@Configuration
+@RequiredArgsConstructor
+public class SecurityConfig {
+
+    private final PermissionRuleUriMapper permissionRuleUriMapper;
+    private final KeycloakJwtAuthenticationConverter keycloakJwtAuthenticationConverter;
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http
+                .csrf(csrf -> csrf.disable())
+                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+                .authorizeHttpRequests(authz -> {
+                    permissionRuleUriMapper.configureAuthorization(authz);
+                    authz.anyRequest().authenticated();
+                })
+                .oauth2ResourceServer(oauth2 -> oauth2.jwt(jwt -> jwt.jwtAuthenticationConverter(keycloakJwtAuthenticationConverter)))
+                .exceptionHandling(this::configureExceptionHandling);
+
+        return http.build();
+    }
+
+
+    private void configureExceptionHandling(ExceptionHandlingConfigurer<HttpSecurity> ex) {
+        ex.authenticationEntryPoint((request, response, authException) -> {
+            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED); // 401
+        }).accessDeniedHandler((request, response, accessDeniedException) -> {
+            response.setStatus(HttpServletResponse.SC_FORBIDDEN); // 403
+        });
+    }
+}
+
